+++ identityhub-service/src/main/java/br/dev/andrestamatto/identityhub/clientapplication/application/BffClientSecretRotator.java
@@ -1,0 +1,6 @@
+package br.dev.andrestamatto.identityhub.clientapplication.application;
+
+public interface BffClientSecretRotator {
+
+    ConfidentialClientSecret rotate(ApplicationClientSnapshot client);
+}