+++ identityhub-service/src/main/java/br/dev/andrestamatto/identityhub/clientapplication/domain/ApplicationClientType.java
@@ -2,5 +2,6 @@ package br.dev.andrestamatto.identityhub.clientapplication.domain;
 
 public enum ApplicationClientType {
     API,
-    SPA
+    SPA,
+    BFF
 }